Commit #3
Added a dynamic prefix handler. Users can use either guild prefix or bot prefix. Commands now include both a group and owner string. Group is to allow for dynamic help command, with the owner bool being used to tell if it is an owner command or not. All owner commands will be non economy related, with a subfolder in economy dedicated to economy related owner commands. Also made the store.ts file, this will house both the shop command and the help command.